+++ ng/ng.h
@@ -4,102 +4,56 @@
 #ifndef NG_H
 #define NG_H
 
-#include <cstddef>
-#include <cstdbool>
-#include <cstdint>
-#include <cstdlib>
-#include <cstdio>
-#include <cstring>
-#include <cerrno>
+#include <stdexcept>
+#include <new>
+#include <iterator>
+#include <vector>
+#include <string>
+#include <sstream>
+#include <iostream>
+#include <fstream>
+#include <filesystem>
 #include <cmath>
+
 #include "SDL.h"
 
-// Core
-// prints "function:line\n"
-#define ng_here(); printf("%s:%d\n",__func__,__LINE__);
+// Engie
+#define NG_HERE __FILE__ << ":" << __LINE__
 
-// prints "function:line: your-print-statement"
-#define ng_debug printf("%s:%d: ",__func__,__LINE__);printf
+namespace ng {
 
-// These are compatible with bool
-enum ngEnumNone { NG_NONE = 0 };
-enum ngEnumError { NG_ERROR = 0, NG_SUCCESS = 1 };
-enum ngEnumTernary { NG_FALSE = 0, NG_EDGE = 1, NG_TRUE = 2 };
+// Use exceptions instead of returning an error code.
+// Throw logic_error when parameters fail checks at the start of a function.
+// Throw runtime_error when a problem occurs during function execution.
+// For example:
+/*
+if (SDL_Init(...) != 0) {
+	throw std::runtime_error(SDL_GetError());
+}
+...
+catch (const std::exception& ex) {
+	std::cout << NG_HERE << ": can't init: " << ex.what();
+}
+*/
+
+// Every code that can be something or nothing, uses ng::None for nothing.
+enum EnumNone { None = 0 };
+
+// ng::Edge represents a fuzzy condition that could be true, false, or both.
+// If you only need true and false, you should use a bool instead.
+// If you need three mutually exclusive states, you should use your own enum instead.
+enum EnumTernary { False = 0, Edge = 1, True = 2 };
+
+void init ();
+void quit ();
+
+} // namespace ng
 
 // String
-// Wrappers around clib's string functions - with boundary checking and
-// memory re-allocation - to (hopefully) avoid undefined behavior.
-/*
-int64_t ng_strlen (const char*);
-int ng_strcmp (const char*, const char*);
-char* ng_strnul (char*); // str beginning with ascii 0 (nul), hence strnul.
-char* ng_strcpy (char*, const char*);
-char* ng_strcat (char*, const char*);
-char* ng_strcatc (char*, char); // calls strcat(s,char[2])
-char* ng_substr (char*, const char*, int64_t, int64_t);
-int64_t ng_strdelim (const char*, const char*, int64_t);
-int64_t ng_strndelim (const char*, const char*, int64_t);
-// ^ first delim/non-delim char from start
-// ^ calls strncpy(s,a+start,len)
-char* ng_qstrcpy (char*, const char*, int64_t);
-char* ng_qstrcat (char*, int64_t, const char*, int64_t);
-char* ng_qstrcatc (char*, int64_t, char);
-char* ng_qsubstr (char*, const char*, int64_t, int64_t, int64_t);
-int64_t ng_qstrdelim (const char*, int64_t, const char*, int64_t);
-int64_t ng_qstrndelim (const char*, int64_t, const char*, int64_t);
-// ^ quick str functions. the 'q' stands for quick.
-bool ng_strchr (const char*, char); // if char in str.
-void ng_strupper (char*);
-void ng_strlower (char*);
-char* ng_atoh (char*, const char*, char);
-// ^ pass an uppercase/lowercase letter to output uppercase/lowercase hex.
-// ^ if the letter is X/x, also prepends "0x".
-double ng_atof (const char*);
-int64_t ng_atoi (const char*);
-bool ng_atob (const char*);
-char* ng_htoa (char*, const char*);
-char* ng_ftoa (char*, const char*, double); // calls sprintf(s,fmt,d)
-char* ng_itoa (char*, const char*, int64_t); // calls sprintf(s,fmt,i)
-char* ng_btoa (char*, bool);
-// more quick str functions
-char* ng_qatoh (char*, const char*, int64_t, char c);
-char* ng_qhtoa (char*, const char*, int64_t);
-*/
+// use <string> and <sstream> instead.
 
 // File
-// Wrappers around clib's file functions - with boundary checking and
-// error checking - to (hopefully) avoid undefined behavior.
-/*
-FILE* ng_fopen (const char*, const char*); // returns NULL on error
-char* ng_fgets (char*, FILE*, int buf); // returns chars read before error/EOF
-// Buf is the number of chars to read at a time.
-// - Larger buffers allow faster reads, but skip data when an error occurs,
-//   and require ftell and fseek to return to where it left off.
-// - Reading 1 char at a time is slow, but will pick up right where it
-//   left off after errors are dealt with.
-int64_t ng_fputs (const char*, FILE*); // returns chars written before error/len
-// ^ writes 1 char at a time, as if by repeated calls to fputc.
-int ng_fputc (char, FILE*);
-FILE* ng_fclose (FILE*); // always returns NULL
-int ng_rename (const char*, const char*);
-int ng_delete (const char*);
-bool ng_exist (const char*, char); // file exists and program/user has permission
-// check for 'r'ead/'w'rite/'+'both access
-*/
-
-// Folders are OS-dependent
-// Linux: int mkdir (const char*, int mode); requires <sys/stat.h>
-// Windows: int mkdir (const char*); requires <windows.h>
-// ng_delete can remove empty folders
-// TODO List files in a folder?
-
-/*
-bool ng_file_exists (const char* file);
-void ng_file_delete (const char* file);
-char* ng_file_read (const char* file);
-void ng_file_write (const char* file, const char* s);
-void ng_file_append (const char* file, const char* s);
-*/
+// use <iostream> and <fstream> and <filesystem> instead.
 
 // Time
 typedef struct ngTime {
@@ -172,25 +126,25 @@ void ng_grid_portal (ngGrid*, const ngRect* src, const ngRect* dest);
 
 void ng_color_init (ngColor*, int r, int g, int b);
 
-int ng_image_init (ngGraphics*, ngImage*, const char* file, const ngColor* key);
+void ng_image_init (ngGraphics*, ngImage*, const char* file, const ngColor* key);
 void ng_image_quit (ngImage*);
-int ng_image_color (ngImage*, const ngColor*);
-int ng_image_alpha (ngImage*, const ngColor*);
+void ng_image_color (ngImage*, const ngColor*);
+void ng_image_alpha (ngImage*, const ngColor*);
 void ng_image_flip (ngImage*, int flip);
 void ng_image_angle (ngImage*, double angle);
 
-int ng_graphics_init (ngGraphics*, const char* title, int w, int h);
+void ng_graphics_init (ngGraphics*, const char* title, int w, int h);
 void ng_graphics_quit (ngGraphics*);
-int ng_graphics_color (ngGraphics*, const ngColor*);
-int ng_graphics_alpha (ngGraphics*, const ngColor*);
-int ng_graphics_clear (ngGraphics*);
+void ng_graphics_color (ngGraphics*, const ngColor*);
+void ng_graphics_alpha (ngGraphics*, const ngColor*);
+void ng_graphics_clear (ngGraphics*);
 void ng_graphics_draw (ngGraphics*);
 
-int ng_draw_image (ngGraphics*, ngImage*, const ngRect*, const ngRect*);
+void ng_draw_image (ngGraphics*, ngImage*, const ngRect*, const ngRect*);
 enum ngEnumRectFill { NG_FRAME = 0, NG_FILL = 1 }; // compatible with bool
-int ng_draw_rect (ngGraphics*, const ngRect*, bool fill);
-int ng_draw_line (ngGraphics*, int x1, int y1, int x2, int y2);
-int ng_draw_point (ngGraphics*, int x, int y);
+void ng_draw_rect (ngGraphics*, const ngRect*, bool fill);
+void ng_draw_line (ngGraphics*, int x1, int y1, int x2, int y2);
+void ng_draw_point (ngGraphics*, int x, int y);
 
 /*
 void ng_graphics_render_text (ngGraphics* g, const ngFrame* frame, int w_frames, int h_frames,
@@ -293,9 +247,9 @@ typedef struct ngAudio {
 	bool playing;
 } ngAudio;
 
-int ng_audio_init (ngAudio*);
+void ng_audio_init (ngAudio*);
 void ng_audio_quit (ngAudio*);
-int ng_clip_init (ngAudio*, ngClip*, const char* file);
+void ng_clip_init (ngAudio*, ngClip*, const char* file);
 void ng_clip_quit (ngClip*);
 void ng_channel_init (ngChannel*);
 void ng_channel_quit (ngChannel*);
@@ -313,7 +267,7 @@ void ng_audio_clear (ngAudio*); // fill queue with silence
 int ng_audio_mix_sample (int a, int b); // add decibels
 void ng_audio_mix_sound (ngAudio*, ngChannel*, int sound); // add clip data to queue
 void ng_audio_mix_channel (ngAudio*, ngChannel*);
-int ng_audio_queue (ngAudio*); // queue audio
+void ng_audio_queue (ngAudio*); // queue audio
 
 // Physics
 typedef struct ngVec {